+++ package.json
@@ -14,10 +14,12 @@
     "react-dom": "^16.8.4",
     "react-final-form": "^4.1.0",
     "react-redux": "^6.0.1",
+    "react-router-dom": "^5.0.0",
     "react-scripts": "2.1.8",
     "redux": "^4.0.1",
     "redux-promise-middleware": "^6.1.0",
-    "redux-thunk": "^2.3.0"
+    "redux-thunk": "^2.3.0",
+    "typescript": "^3.4.1"
   },
   "scripts": {
     "start": "react-scripts start",
