+++ src/actions/index.js
@@ -1,10 +1,15 @@
-import { GET_FILMS, RESET_SELECTED_FILM, SAVE_FILM, SELECT_FILM } from "../types";
+import {GET_FILM, GET_FILMS, RESET_SELECTED_FILM, SAVE_FILM, SELECT_FILM} from "../types";
 import axios from 'axios';
 
 
 export const getFilms = () => ({
   type: GET_FILMS,
   payload: axios.get('https://www.omdbapi.com/?apikey=792db8bc&s=marvel')
+})
+
+export const getFilm = (id) => ({
+  type: GET_FILM,
+  payload: id
 })
 
 export const selectFilm = selectedFilm => ({
