+++ src/index.js
@@ -1,13 +1,14 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './components/App';
-import { Provider } from "react-redux";
-import { applyMiddleware, compose, createStore } from "redux";
+import {Provider} from "react-redux";
+import {applyMiddleware, compose, createStore} from "redux";
 import reducer from "./reducers";
 import MuiThemeProvider from "@material-ui/core/styles/MuiThemeProvider";
-import { createMuiTheme } from "@material-ui/core";
+import {createMuiTheme} from "@material-ui/core";
 import thunk from 'redux-thunk';
 import promise from 'redux-promise-middleware'
+import {BrowserRouter} from "react-router-dom";
 
 const theme = createMuiTheme({
   palette: {type: 'dark'},
@@ -24,7 +25,9 @@ const store = createStore(
 ReactDOM.render(
   <Provider store={store}>
     <MuiThemeProvider theme={theme}>
-      <App/>
+      <BrowserRouter>
+        <App/>
+      </BrowserRouter>
     </MuiThemeProvider>
   </Provider>,
   document.getElementById('root')
