+++ src/types/index.js
@@ -2,6 +2,7 @@ export const GET_FILMS = 'GET_FILMS';
 export const GET_FILMS_FULFILLED = 'GET_FILMS_FULFILLED';
 export const GET_FILMS_REJECTED = 'GET_FILMS_REJECTED';
 
+export const GET_FILM = 'GET_FILM';
 export const RESET_SELECTED_FILM = 'RESET_SELECTED_FILM';
 export const SELECT_FILM = 'SELECT_FILM';
 export const SAVE_FILM = 'SAVE_FILM';
